+++ tools/generate_test_data.py
@@ -13,7 +13,7 @@ import uvloop
 from cryptofeed.feedhandler import FeedHandler
 from cryptofeed.exchanges import EXCHANGE_MAP
 from cryptofeed.raw_data_collection import AsyncFileCallback
-from cryptofeed.defines import BINANCE, BINANCE_FUTURES, BINANCE_US, BITFINEX, L2_BOOK, TRADES, TICKER, CANDLES, EXX
+from cryptofeed.defines import BINANCE, BINANCE_FUTURES, BINANCE_US, BINANCE_TR, BITFINEX, L2_BOOK, TRADES, TICKER, CANDLES, EXX
 from check_raw_dump import main as check_dump
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
@@ -43,7 +43,7 @@ def main(only_exchange=None):
         info = exchange.info()
         channels = list(set.intersection(set(info['channels']['websocket']), set([L2_BOOK, TRADES, TICKER, CANDLES])))
         sample_size = 10
-        if exch_str in (BINANCE_US, BINANCE):
+        if exch_str in (BINANCE_US, BINANCE_TR, BINANCE):
             # books of size 5000 count significantly against rate limits
             sample_size = 4
         while True: