+++ CHANGES.md
@@ -8,6 +8,7 @@
  * Feature: Use backend Queue for Kafka
  * Feature: Add support for storing book snapshots in Redis as key-value
  * Update: Switch from unmaintained aioredis to redis-py
+ * Bugfix: Correct value for Crypto.com Ask price
 
 ### 2.3.1 (2022-10-31)
  * Bugfix: timestamp not reset correctly on reconnect